+++ CHANGELOG.rst
@@ -1,12 +1,13 @@
 Brewtils Changelog
 ==================
 
-TBD
+3.34.0
 ------
 TBD
 
 - Added RabbitMQ remote disconnect logic for Plugin to shutdown if RabbitMQ closes the channel connection or rejects Topic subscription (#589) 
 - Added Plugin re-register logic for if Instance is unknown by Beer-Garden (#589) 
+
 
 3.33.0
 ------